backend/requirements.txt
python-dotenv==1.0.0
openai==1.12.0
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
Flask==3.0.2
Flask-CORS==4.0.0
Werkzeug==3.0.1
